backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.2
librosa==0.10.1
soundfile==0.12.1
requests==2.31.0
python-dotenv==1.0.0
tensorflow==2.15.0

